requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.0.7
django-crispy-forms==2.2
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
